+++ c-examples/cdh-monolithic/support/cdh.c
@@ -23,18 +23,43 @@
 
 // Functions required by TAB
 
-// Toy example that checks that data is strictly in ascending order
+// This example implementation of handle_common_data checks whether the bytes
+// are strictly increasing, i.e. each subsequent byte is strictly greater than
+// the previous byte
 int handle_common_data(common_data_t common_data_buff_i) {
-  uint8_t prev_byte;
   int strictly_increasing = 1;
+  uint8_t prev_byte = common_data_buff_i.data[0];
   for(size_t i=1; i<common_data_buff_i.end_index; i++) {
-    prev_byte = common_data_buff_i.data[i-1];
     if(prev_byte>=common_data_buff_i.data[i]) {
       strictly_increasing = 0;
       i = common_data_buff_i.end_index;
+    } else {
+      prev_byte = common_data_buff_i.data[i];
     }
   }
   return strictly_increasing;
+}
+
+// This example implementation of handle_bootloader_erase erases all application
+// programs
+int handle_bootloader_erase(void){
+  flash_unlock();
+  for(size_t subpage_id=0; subpage_id<255; subpage_id++) {
+    // subpage_id==0x00 writes to APP_ADDR==0x08008000 i.e. start of page 16
+    // So subpage_id==0x10 writes to addr 0x08008800 i.e. start of page 17 etc
+    // Need to erase page once before writing inside of it
+    if((subpage_id*BYTES_PER_CMD)%BYTES_PER_PAGE==0) {
+      flash_erase_page(16+(subpage_id*BYTES_PER_CMD)/BYTES_PER_PAGE);
+      flash_clear_status_flags();
+    }
+  }
+  flash_lock();
+  return 1;
+}
+
+// This example implementation of bootloader_active always returns true
+int bootloader_active(void) {
+  return 1;
 }
 
 // Board initialization functions