+++ c-examples/cdh-monolithic/support/cdh.h
@@ -2,7 +2,7 @@
 // CDH board support header file
 //
 // Written by Bradley Denby
-// Other contributors: None
+// Other contributors: Chad Taylor
 //
 // See the top-level LICENSE file for the license.
 
@@ -10,10 +10,17 @@
 #define CDH_H
 
 // TAB header
-#include <tab.h> // rx_cmd_buff_t, tx_cmd_buff_t, common_data_t
+#include <tab.h> // common_data_t, rx_cmd_buff_t, tx_cmd_buff_t
+
+//// Byte counts
+#define BYTES_PER_CMD  ((uint32_t)128)
+#define BYTES_PER_PAGE ((uint32_t)2048)
 
 // Functions required by TAB
+
 int handle_common_data(common_data_t common_data_buff_i);
+int handle_bootloader_erase(void);
+int bootloader_active(void);
 
 // Board initialization functions